Build top-nav that sits correctly
Before this commit, the Google logo was not sitting correctly
in line with the other elements in the top div.
With this commit, I discovered the issue was
with the vertical-align property not being set to a value which
would align the divs with .top.
I will now go back to working on the main branch
and neaten up the code there to make it work.