Style top-nav tab links correctly
The top nav links had regular link hover properties.
The Tools button specifically is different in the original.
With this commit, the Tools button is adjusted and, thus,
the right side positioning adjusted to compensate.
This allows for its proper styling.
All other anchor elements in the tabs section are also styled
as per the original, with this commit.