Tweak and finalize site-spans in results
The positioning of the green site-spans in results paragraphs
finishes with a downward facing arrow and also was positioned better
in the original with regard to the heading and the rest
of each paragraph.
With this commit, arrows are added to each span
using the ::after pseudo element on each site-span.
Spacing of the spans is also corrected.
All that remains unfinished on the page now is paragraph line-spacing.