Tweak spacing of results paragraph elements
With this commit, further adjustments are made to the spacing
of various paragraph elements, to try to achieve a spacing similar
to the original.
However, the spacing is still not a perfect replica.
For now, since the rest of the page is matching though,
I will leave this project and move on.